+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2022-06-08
+### Added
+- Command to download dataset
+- Make instruction to download dataset
+
 ## [2.3.3] - 2022-06-08
 ### Added
 - Handling of SIGTERM on workers